+++ src/pages/List/components/Modal/CreditAlertModal/index.jsx
@@ -1,22 +1,12 @@
 import closeIcon from '@/assets/icons/ic_close.svg';
 import creditImg from '@/assets/images/img_diamond.png';
 import style from './styles.module.scss';
-import { useRef } from 'react';
+import ModalBackground from '../components/ModalBackground';
 
 const CreditAlertModal = ({ isModalOpen, closeModal }) => {
-  const modalBackgroundRef = useRef();
   return (
     <>
-      <div
-        style={{ display: isModalOpen ? 'block' : 'none' }}
-        className={style['modal-background']}
-        ref={modalBackgroundRef}
-        onClick={(e) => {
-          if (e.target === modalBackgroundRef.current) {
-            closeModal();
-          }
-        }}
-      >
+      <ModalBackground isModalOpen={isModalOpen} closeModal={closeModal}>
         <div
           style={{ display: isModalOpen ? 'block' : 'none' }}
           className={style['container']}
@@ -37,7 +27,7 @@ const CreditAlertModal = ({ isModalOpen, closeModal }) => {
             <button>확인</button>
           </div>
         </div>
-      </div>
+      </ModalBackground>
     </>
   );
 };
