+++ src/pages/List/components/Modal/CreditAlertModal/styles.module.scss
@@ -1,14 +1,5 @@
 @import '@/styles/import.scss';
 
-.modal-background {
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  background-color: rgba(0, 0, 0, 0.65);
-}
-
 .container {
   position: absolute;
   top: 50%;
@@ -18,6 +9,7 @@
   height: 331px;
   background-color: #181d26;
   border-radius: 12px;
+  box-shadow: 0px 4px 4px $tj-color-white-aplha-10;
 
   > .header {
     text-align: right;
@@ -35,8 +27,7 @@
   }
 
   > .main {
-    display: flex;
-    flex-direction: column;
+    @include Flex(column, flex-start, stretch, 0);
 
     > img {
       width: 150px;
@@ -47,30 +38,29 @@
     > span {
       font: 500 16px/26px Pretendard;
       text-align: center;
-      color: #ffffff;
+      color: $tj-color-white;
       margin: 0px 0px 31px;
 
       > .strong {
-        color: #f96d69;
+        color: $tj-color-orange;
       }
     }
   }
 
   > .footer {
-    display: flex;
-    justify-content: center;
+    @include Flex(row, center, stretch, 0);
     > button {
       cursor: pointer;
       width: 295px;
       height: 45px;
       margin: 0 auto;
       border-radius: 3px;
-      background: linear-gradient(90deg, #f86f65, #fe5493);
+      background: $tj-gradient-orange-to-pink;
       font: 700 14px/26px Pretendard;
-      color: #ffffff;
+      color: $tj-color-white;
 
       &:hover {
-        background: linear-gradient(270deg, #f86f65, #fe5493);
+        background: $tj-gradient-orange-to-pink-reverse;
       }
     }
   }
