+++ src/pages/List/components/Modal/IdolDonationModal/styles.module.scss
@@ -1,14 +1,5 @@
 @import '@/styles/import.scss';
 
-.modal_background {
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  background-color: rgba(0, 0, 0, 0.65);
-}
-
 .container {
   position: absolute;
   top: 50%;
@@ -19,16 +10,15 @@
   border-radius: 8px;
   background-color: #181d26;
   padding: 24px 19px 32px;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  box-shadow: 0px 4px 4px $tj-color-white-aplha-10;
 
   > .header {
-    display: flex;
-    justify-content: space-between;
+    @include Flex(row, space-between, stretch, 0);
     margin-bottom: 24px;
 
     > h2 {
       font: 600 18px/21.48px Pretendard;
-      color: #f7f7f8;
+      color: $tj-color-white;
     }
 
     > button {
@@ -44,9 +34,7 @@
 
   > .main {
     > .card {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
+      @include Flex(column, flex-start, center, 0);
 
       > img {
         margin: 0 auto;
@@ -61,12 +49,12 @@
 
       > .text-wrapper > span {
         font: 400 12px/18px Pretendard;
-        color: rgba(255, 255, 255, 0.4);
+        color: $tj-color-white-aplha-40;
       }
 
       > .text-wrapper > p {
         font: 500 14px/16.71px Pretendard;
-        color: #f7f7f8;
+        color: $tj-color-white;
         margin-top: 6px;
       }
     }
@@ -78,10 +66,10 @@
         width: 290px;
         height: 58px;
         border-radius: 8px;
-        border: 1px solid #ffffff;
+        border: 1px solid $tj-color-white;
         background-color: #272f3d;
         font: 700 20px/26px Pretendard;
-        color: #ffffff;
+        color: $tj-color-white;
         padding: 16px;
       }
 
@@ -102,9 +90,7 @@
 }
 
 .footer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
+  @include Flex(row, center, center, 0);
   margin-top: 24px;
 
   > button {
@@ -112,12 +98,12 @@
     width: 295px;
     height: 42px;
     border-radius: 3px;
-    background: linear-gradient(90deg, #f86f65, #fe5493);
+    background: $tj-gradient-orange-to-pink;
     font: 700 14px/26px Pretendard;
-    color: #ffffff;
+    color: $tj-color-white;
 
     &:hover {
-      background: linear-gradient(270deg, #f86f65, #fe5493);
+      background: $tj-gradient-orange-to-pink-reverse;
     }
   }
 }
