+++ src/pages/List/components/Modal/components/ModalBackground/styles.module.scss
@@ -1,0 +1,8 @@
+.modal-background {
+  position: fixed;
+  width: 100vw;
+  height: 100vh;
+  top: 0;
+  left: 0;
+  background-color: rgba(0, 0, 0, 0.65);
+}
